runtime: add shared_memo
nod copies it's slots when notifying a signal. When
using a memo inside a selector, the memo could remember
only it's initial call. Because afterwards nod was
copying the memo and the memo updated it's state
on the copy. For this we introduce a shared memo
which is wrapped in shared_ptr so we can easily
copy it, but the memo would update always the shared
state.