runtime: remove enum_io_processor
This didn't worked with midi assigned parameters. In the
case when the parent component is not inserted into
the graph, the midi assignment code created it's own
parameter processor, but since it couldn't figure out
the exact enum type, it just create an int parameter
processor. Which might be not compatible with parent
component, which expects an enum_io_processor. If
an enum type is really required in the graph, we
can always insert a converter processor.