runtime: allow midi mapping of aux_send active parameter
Therefore we also need to export the mapping into session file.
If a paramter is not midi assignable we have now a dedicated
flag, instead of relying on the audio_graph_affecting flag.
The midi assignment to active parameter gets disabled if
it would potentially introduce a feedback loop into the graph.
The update is not instant on midi arrival but requires
a graph rebuild.