runtime: automatically unload ladspa instances
Previously we only unloaded when a specific action was triggered,
e.g. delete_fx_module or delete_mixer_channel. But thats too
restricted and error prone. Instead we keep a copy of loaded
instances in the middleware and compare with the one in the state
after every action. The comparison and copy is cheap because
we switched to a boxed map.